third_party/WebKit/LayoutTests/gamepad/gamepad-api.html

<!DOCTYPE html>
<body>
<script src="../fast/js/resources/js-test-pre.js"></script>
<script>
    shouldBeDefined("navigator.webkitGetGamepads");
</script>
<p>Make sure the main polling access point exists on navigator.</p>
</body>
